+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Improvement: Rename all test cases to `*.test.js` files;
 - Improvement: Move mocha options to `mocha.opts` files in `test` folder;
 - Improvement: Optimize `passport.js` configuration file;
+- Improvement: Move ping from Auth to `PingController`. Into this controller can be added more checks of server status;
 - Typo: Rename `JWT` header auth to `Bearer`;
 - Typo: Remove auth by username. Local strategy works only with email now;
 - Typo: Rename `fix-deps` script to `check-deps`;